refactor: 마인드맵 기능 분리
